Reduced the number of checks
Swapped the conditions in the “find_book_title” function which checks if the character is a space and if the space is right before the match. This is so that the bot checks for a space first before checking if the space is right before the match, which means the second check is run at most twice, compared to the other way around where both checks would be run for every single character, minus the one time when the index is one less than the match_index.