Actually fixed the split message function.
The previous implementation of the split message function was truly a hacky mess, it relied on getting the next part of the message with the end index of the part being just 1 more than the maximum length. Then, the line after that was checking if the length of the part was more than the maximum length to determine if that part of the message was the final part of the message so that the bot will break the loop and append that final part to the list of messages to send. The previous commit that "fixed" the function, 5a1c020, supposedly "fixed" the off by one error, but it really only added a max_len variable so the function is more flexible, and a massive bug that renders the split message function useless, as it will only send two messages at most, even when the text required more than two messages. The hack wasn't obvious and the comment above it seems like the hack was completely unintentional. What a stroke of luck. Anyway, it is finally fixed for real this time.