Used enumerate instead of range and len.
Completed the comment on line 885 that wasn't complete and updated the for loop used to send the messages to use enumerate instead of range and len.

This commit is also to explain the changes to the check_backticks and the iterate_text function in the previous commit.

The iterate_text function just returns the first index + the result of the check_backticks function now. Everything is handled within the check_backticks function so that it is much easier to keep track and there's no need to do complicated math to figure out what the iterate_text function is returning.