Cleaned up the regexes.
Made all the regular expressions more concise by using the curly braces construct instead of writing the token to match again and adding a question mark (?) behind it to signify that it is optional.

Changed the comment on line 726 to reflect the line of code below accurately (Esther 5 -> Greek Esther 5).

Made all the equality checks to use tuples instead of lists to save on memory and be consistent with the rest of the code, as some were using tuples and others were using lists.