Now the verse of the day follows the saved version
With this change, the get_verse_of_the_day function has been changed to be asynchronous, so all of the functions calling the function now have to use asyncio.run to call the get_verse_of_the_day function.

The verse of the day will default to NIV if the passage isn't available in a specific version (DRA in particular doesn't have Psalms).

Also edited the help command to document that the verse of the day will follow the bible version saved.

The "chats_version" key in the database now stores a dictionary instead of a list, which makes it easier to work with. The increase in storage size also isn't much, which is great. A lot of functions that interact with the "chats_version" key in the database were rewritten to support the new data structure.

Removed the obtain_version function as it is obsolete. The function to get the version is now a one-liner in the get_version function.

Also added more spacing between the function definitions for greater readability,

Also cleaned up the get_webpages function by removing unnecessary lines of code.