Edited the get_verse_of_the_day function to be synchronous
This is so that I can handle the error and get it to try again instead of just erroring out and killing the long-running thread to send the verse of the day. After all, the speed of sending out the verse of the day doesn't really matter as it's not a request time thing.