Updated the error handling for the get_webpages function
The previous way to handle any errors that got returned by httpx when requesting the webpages was to keep awaiting the asyncio.gather function, which doesn't work. So now the get_webpages function has similar error handling to the send_verse_of_the_day function, where the asyncio.gather call will return exceptions as values in the list and the bot will handle them by iterating over the list and checking if the item is an exception. The bot will then keep requesting the webpage until it's successful and replace the exception with the successful response to the list.

Also updated the packages used.