Made the verse of the day display properly
With that, a new module called utils.py was added and it contains the to_sups function to turn text into the superscript version. The implementation is taken straight out of the verse_match class.

Updated the get_verse_of_the_day function to make sure that all the <br> tags are converted to a new line character and turned all of the text in <sup> tags to superscript.

Also updated the packages used.