Replaced the backticks in the verse of the day with a different one
This is so that the backticks that are present in the bible verse do not mess with the backticks used for markdown formatting when sending the message to the user.

Also updated the function used to replace the backticks in the verse match class to use the normal string.replace instead of a regular expression substitution, so it should be faster and more efficient now. The comment for the line of code that replaces the backticks in the verse match class has also been updated to match the comment for the same line of code in the get_verse_of_the_day function.

Also provided a default value of None to the self.convos.pop function in the next step handler so the pop function will not throw an error when there is no item to remove from the dictionary, which happens when the register_next_step_handler function hasn't been called before the clear_step_handler function. This occurs when the next step is immediately run, like when the version is provided in the message containing the /setversion command, which means the bot doesn't have to register a next step handler to prompt the user for the version, but can instead immediately set the version straightaway by calling the function for the next step.