Removed random characters from the verse of the day message
This should fix the errors with sending the verse of the day message in the MSG version of the bible since the asterisks were causing parsing issues in markdown.

Also added more information to the comment in the verse_match.pyx file, above the line that removes random characters, to state that the line is removing asterisks from the MSG version of the bible.

I feel like I should have used HTML parsing instead of markdown, it's such a pain to be replacing all the markdown reserved characters with something else or removing them entirely.